Move files for less confusing source
The code that is made available in the templates module of projects
using ructe is not moved into the templates module of ructe.

Also, the files containg that code is now handled as proper
submodules, both ructe and in generated code.
