+++ src/templates/utils_warp02.rs
@@ -1,4 +1,5 @@
 use mime::TEXT_HTML_UTF_8;
+use std::io;
 use warp::http::{header::CONTENT_TYPE, response::Builder};
 use warp::{reject::custom, reject::Reject, reply::Response, Rejection};
